+++ crud-api/package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "build": "webpack --config webpack.config.ts",
     "start:dev": "cross-env NODE_ENV=development ts-node-dev --respawn src/index.ts",
-    "start:prod": "npm run build && cross-env NODE_ENV=production node dist/index.js"
+    "start:prod": "npm run build && cross-env NODE_ENV=production node dist/index.js",
+    "test": "jest"
   },
   "author": "playoffthecuff",
   "license": "ISC",
@@ -29,7 +30,9 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.13.0",
+    "@types/jest": "^29.5.13",
     "@types/node": "^22.7.7",
+    "@types/supertest": "^6.0.2",
     "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^8.10.0",
     "@typescript-eslint/parser": "^8.10.0",
@@ -39,7 +42,10 @@
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-prettier": "^5.2.1",
     "globals": "^15.11.0",
+    "jest": "^29.7.0",
     "prettier": "^3.3.3",
+    "supertest": "^7.0.0",
+    "ts-jest": "^29.2.5",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
